CodeMirror view/model split, no longer IFluidHTMLView
