+++ examples/apps/spaces/src/fluid-object/spacesItemMap.ts
@@ -11,7 +11,7 @@ import {
     IFluidDataStoreContext,
 } from "@fluidframework/runtime-definitions";
 import { ReactViewAdapter } from "@fluidframework/view-adapters";
-import { SmdeFactory } from "@fluid-example/codemirror";
+import { CodeMirrorComponent, CodeMirrorView, SmdeFactory } from "@fluid-example/codemirror";
 import { CollaborativeText, CollaborativeTextView } from "@fluid-example/collaborative-textarea";
 import { Coordinate } from "@fluid-example/multiview-coordinate-model";
 import { SliderCoordinateView } from "@fluid-example/multiview-slider-coordinate-view";
@@ -45,22 +45,31 @@ const getAdaptedViewForSingleHandleItem = async (serializableObject: ISingleHand
     return React.createElement(ReactViewAdapter, { view: component });
 };
 
-const getSliderCoordinateView = async (serializableObject: ISingleHandleItem) => {
-    const handle = serializableObject.handle as IFluidHandle<Coordinate>;
-    const model = await handle.get();
-    return React.createElement(SliderCoordinateView, { label: "Coordinate", model });
-};
-
 const getClickerView = async (serializableObject: ISingleHandleItem) => {
     const handle = serializableObject.handle as IFluidHandle<Clicker>;
     const clicker = await handle.get();
     return React.createElement(ClickerReactView, { clicker });
 };
 
+const getCodeMirrorView = async (serializableObject: ISingleHandleItem) => {
+    const handle = serializableObject.handle as IFluidHandle<CodeMirrorComponent>;
+    const codeMirror = await handle.get();
+    return React.createElement(
+        ReactViewAdapter,
+        { view: new CodeMirrorView(codeMirror.text, codeMirror.presenceManager) },
+    );
+};
+
 const getCollaborativeTextView = async (serializableObject: ISingleHandleItem) => {
     const handle = serializableObject.handle as IFluidHandle<CollaborativeText>;
     const collaborativeText = await handle.get();
     return React.createElement(CollaborativeTextView, { text: collaborativeText.text });
+};
+
+const getSliderCoordinateView = async (serializableObject: ISingleHandleItem) => {
+    const handle = serializableObject.handle as IFluidHandle<Coordinate>;
+    const model = await handle.get();
+    return React.createElement(SliderCoordinateView, { label: "Coordinate", model });
 };
 
 /**
@@ -84,7 +93,7 @@ const clickerItemEntry: ISpacesItemEntry<ISingleHandleItem> = {
 
 const codemirrorItemEntry: ISpacesItemEntry<ISingleHandleItem> = {
     create: createSingleHandleItem(codeMirrorFactory),
-    getView: getAdaptedViewForSingleHandleItem,
+    getView: getCodeMirrorView,
     friendlyName: "Code",
     fabricIconName: "Code",
 };
