+++ examples/data-objects/codemirror/package.json
@@ -38,6 +38,7 @@
     "@fluidframework/aqueduct": "^0.58.0",
     "@fluidframework/container-definitions": "^0.46.0-0",
     "@fluidframework/container-runtime": "^0.58.0",
+    "@fluidframework/container-runtime-definitions": "^0.58.0",
     "@fluidframework/core-interfaces": "^0.42.0",
     "@fluidframework/datastore": "^0.58.0",
     "@fluidframework/datastore-definitions": "^0.58.0",
@@ -47,6 +48,7 @@
     "@fluidframework/runtime-definitions": "^0.58.0",
     "@fluidframework/runtime-utils": "^0.58.0",
     "@fluidframework/sequence": "^0.58.0",
+    "@fluidframework/view-adapters": "^0.58.0",
     "@fluidframework/view-interfaces": "^0.58.0",
     "@types/codemirror": "0.0.76",
     "codemirror": "^5.48.4"
